solutions/typescript/2019/22/src/p1.spec.ts
import { read } from '@alexaegis/advent-of-code-lib';
import { describe, expect, it } from 'vitest';
import packageJson from '../package.json' assert { type: 'json' };
import { runner } from './p1.js';

describe('2019 - Day 22 - Part One', () => {
	it('should solve the input', async () => {
		const input = await read(packageJson.aoc.year, packageJson.aoc.day)();
		expect(runner()(input.input)).to.equal(6831);
	});

	it('should solve for the first example', async () => {
		const input = await read(packageJson.aoc.year, packageJson.aoc.day, 'example.1.txt')();
		expect(await runner()(input.input)).to.equal(4126);
	});

	it('should solve for the second example', async () => {
		const input = await read(packageJson.aoc.year, packageJson.aoc.day, 'example.2.txt')();
		expect(await runner()(input.input)).to.equal(5922);
	});

	it('should solve for the third example', async () => {
		const input = await read(packageJson.aoc.year, packageJson.aoc.day, 'example.3.txt')();
		expect(await runner()(input.input)).to.equal(7115);
	});

	it('should solve for the fourth example', async () => {
		const input = await read(packageJson.aoc.year, packageJson.aoc.day, 'example.4.txt')();
		expect(await runner()(input.input)).to.equal(1219);
	});
});
